+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- **BREAKING**: Update the default ΔT estimator from 2015 onwards with the [2026 polynomial adaptation](https://klaus.brunners.name/posts/delta-t-polynomials/). This changes estimated values and solar calculations that use them.
+
 ## [0.5.3] - 2026-08-12
 
 ### Changed